feat: Allow empty host on `WithRoute` (#67)
Envoy requires unique domains for all virtual hosts within the same
route configuration. This was causing issues with multiple services
registering under the same domain/host through fuse. This PR removes the
requirement to specify a host when registering routes and #65 updates
fuse to add routes with no defined host to a default virtual host.

Oh and I also added a `Ping()` call to the `repositories/postgres/bun`
client's `Open()` so that the database connection is actually verified
on boot.